# docs/api/openapi.yaml
openapi: 3.1.0
info:
  description: "달리모 서버 API. 응답은 { success, data, error, timestamp } (7.1장), 오류 코드는\
    \ 27.1장. 명세 41~45장과 다른 곳은 frontend/docs/api/MOCK-CONTRACT-CHECK.md"
  title: DALLIMO API
  version: v1
servers:
- description: "지금 연 주소의 서버 (배포: https://dallimo.gamjabox.cloud)"
  url: /
paths:
...
  /api/v1/admin/errors:
    get:
      operationId: errors
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      - in: query
        name: days
        required: false
        schema:
          type: integer
          format: int32
          default: 7
          maximum: 30
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseErrors"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
...
  /api/v1/admin/monitoring:
    get:
      operationId: monitoring
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseMonitoring"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
  /api/v1/admin/notices:
    get:
      operationId: list_3
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListNotice"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
    post:
      operationId: send
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseNotice"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
  /api/v1/admin/notices/audience:
    get:
      operationId: audience
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      - in: query
        name: target
        required: false
        schema:
          type: string
          default: ALL
          enum:
          - ALL
          - IOS
          - ANDROID
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseAudienceView"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
  /api/v1/admin/notices/test:
    post:
      operationId: test
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseTestResult"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
  /api/v1/admin/notices/{id}:
    get:
      operationId: get_3
      parameters:
      - in: header
        name: X-Admin-Key
        required: false
        schema:
          type: string
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseNotice"
          description: OK
      security:
      - adminKey: []
      - bearerAuth: []
      tags:
      - 관리 · 모니터링 · 공지
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        details: {}
        message:
          type: string
...
    ApiResponseAudienceView:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AudienceView"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseErrors:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Errors"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListNotice:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notice"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseMonitoring:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Monitoring"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseNotice:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Notice"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    ApiResponseTestResult:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/TestResult"
        error:
          $ref: "#/components/schemas/ApiError"
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
...
    AudienceView:
      type: object
      properties:
        devices:
          type: integer
          format: int32
        quietHours:
          type: boolean
        users:
          type: integer
          format: int32
...
    Check:
      type: object
      properties:
        error:
          type: string
        ms:
          type: integer
          format: int64
        ok:
          type: boolean
...
    Entry:
      type: object
      properties:
        action:
          type: string
        actor:
          type: string
        actorName:
          type: string
        createdAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        reason:
          type: string
...
    Errors:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        last24h:
          type: integer
          format: int32
        recent:
          type: array
          items:
            $ref: "#/components/schemas/Entry"
...
    Group:
      type: object
      properties:
        count:
          type: integer
          format: int32
        exception:
          type: string
        lastAt:
          type: string
          format: date-time
        lastMessage:
          type: string
        lastPath:
          type: string
        location:
          type: string
...
    KeySummary:
      type: object
      properties:
        errors:
          type: integer
          format: int64
        method:
          type: string
        name:
          type: string
        p95Ms:
          type: integer
          format: int64
        pattern:
          type: string
        requests:
          type: integer
          format: int64
...
    Minute:
      type: object
      properties:
        at:
          type: string
          format: date-time
        errors:
          type: integer
          format: int64
        p95Ms:
          type: integer
          format: int64
        requests:
          type: integer
          format: int64
...
    Monitoring:
      type: object
      properties:
        api:
          $ref: "#/components/schemas/Snapshot"
        at:
          type: string
          format: date-time
        server:
          $ref: "#/components/schemas/Server"
        today:
          $ref: "#/components/schemas/Today"
...
    Notice:
      type: object
      properties:
        actor:
          type: string
        actorName:
          type: string
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        link:
          type: string
        pushFailed:
          type: integer
          format: int32
        pushOk:
          type: integer
          format: int32
        pushTokens:
          type: integer
          format: int32
        status:
          type: string
        target:
          type: string
          enum:
          - ALL
          - IOS
          - ANDROID
        targetUsers:
          type: integer
          format: int32
        title:
          type: string
        tokensRemoved:
          type: integer
          format: int32
...
    SendRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 500
          minLength: 0
        expectedUsers:
          type: integer
          format: int32
          minimum: 1
        link:
          type: string
          pattern: "^/[A-Za-z0-9/_-]{0,199}$"
        target:
          type: string
          enum:
          - ALL
          - IOS
          - ANDROID
        title:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - body
      - expectedUsers
      - target
      - title
    Server:
      type: object
      properties:
        db:
          $ref: "#/components/schemas/Check"
        heapMaxMb:
          type: integer
          format: int64
        heapUsedMb:
          type: integer
          format: int64
        javaVersion:
          type: string
        redis:
          $ref: "#/components/schemas/Check"
        startedAt:
          type: string
          format: date-time
        threads:
          type: integer
          format: int32
        uptimeSec:
          type: integer
          format: int64
...
    Snapshot:
      type: object
      properties:
        keyApis:
          type: array
          items:
            $ref: "#/components/schemas/KeySummary"
        minutes:
          type: array
          items:
            $ref: "#/components/schemas/Minute"
        since:
          type: string
          format: date-time
        total:
          $ref: "#/components/schemas/Summary"
...
    Summary:
      type: object
      properties:
        avgBpm:
          type: integer
          format: int32
        maxBpm:
          type: integer
          format: int32
        sampleCount:
          type: integer
          format: int32
...
    TestRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 500
          minLength: 0
        link:
          type: string
          pattern: "^/[A-Za-z0-9/_-]{0,199}$"
        title:
          type: string
          maxLength: 100
          minLength: 0
        userId:
          type: integer
          format: int64
      required:
      - body
      - title
      - userId
    TestResult:
      type: object
      properties:
        devices:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        ok:
          type: integer
          format: int32
    Today:
      type: object
      properties:
        activeUsers:
          type: integer
          format: int32
        finishedRuns:
          type: integer
          format: int32
        pendingReports:
          type: integer
          format: int32
        serverErrors:
          type: integer
          format: int32
        signups:
          type: integer
          format: int32
...
  securitySchemes:
    adminKey:
      description: 관리 API 키 (ADMIN_API_KEY)
      in: header
      name: X-Admin-Key
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      description: Access Token (14.1장)
      scheme: bearer
      type: http